Fixed some errors and warnings regarding arrays starting from -1, still have problems with pkg
The problems regarding arrays starting from -1 are becuase the initial parameter values were for instance 0, and then when declaring a variable it is (DATA_WIDTH -1) and DATA_WIDTH is 0. That creates problems for yosys where it gives back an error.

Overall the problem seems to be the way the systemverilog plugin interprets the package, because it declares wires that are part of other modules into the package module for no reason, and later on doesn't see the parameters defined in the opl3 package.

Aswell as errors regarding unelaborated modules for some reason.