Updated Amaranth Blinky example
Problem is currently in _gatemate.py file in the file_templates section. The file that generates the .ccf constraint file has a problem reading the directions of the attributes specified in gatemate_a1_evb.py.
So far I have hardcoded the .ccf file in this Blinky Example and it works fine but it needs to be fixed.
I have asked on the Amaranth Discussion on Github for help.